feat: creating toast alert to notifications module
