+++ packages/notifications/package.json
@@ -31,7 +31,8 @@
   "license": "UNLICENSED",
   "dependencies": {
     "@ttoss/config": "^1.9.1",
-    "@ttoss/ui": "^1.9.1"
+    "@ttoss/ui": "^1.9.1",
+    "react-toastify": "^8.2.0"
   },
   "peerDependencies": {
     "react": ">=17.0.2"
