+++ packages/notifications/tsup.config.ts
@@ -1,7 +1,7 @@
 import { tsupConfig } from '@ttoss/config';
 
 export const tsup = tsupConfig({
-  entry: ['src/index.tsx'],
+  entry: ['src/index.ts'],
   minify: false,
   inject: ['./tsup.inject.js'],
 });
